Moved user, posts, comments migration files into one php file. Moved user and post seeders into one php file.
