+++ database/users_seeder.php
@@ -2,9 +2,18 @@
 require 'joblister_db_connect.php';
 require 'db_connect.php';
 // First, add a query to delete all the records from the table
-
-$truncate = 'TRUNCATE users';
+//CAN'T TRUNC A TABLE WITH A FOREIGN KEY
+$truncate = 'DELETE FROM posts';
 $dbc->exec($truncate);
+
+$truncate = 'DELETE FROM users';
+$dbc->exec($truncate);
+
+
+
+
+
+
 
 $Users = [
     ['first_name' => 'Don', 'last_name' => 'Moore','user_name' => 'donny71','email' => 'dmoore@jlister',
@@ -26,5 +35,22 @@ foreach ($Users as $user) {
     $stmt->bindValue(':password', $user['password'], PDO::PARAM_STR);
     $stmt->execute();
 }
+
+
+$posts = [
+    ['user_id' => 1,'title' => 'Javascript developer Needed in Austin!', 'content' => 'We are a new company looking for enthusiastic programmers to help us grow. We have plenty of work, but not enough programmers. Contact us at 555-5555.','date' => '2016-03-31 03:33:33'],
+
+    
+];
+$stmt = $dbc->prepare('INSERT INTO posts (user_id, title, content, date)
+    VALUES (:user_id, :title, :content,:date)');
+foreach ($posts as $post) {
+    $stmt->bindValue(':user_id', $post['user_id'], PDO::PARAM_INT);
+    $stmt->bindValue(':title', $post['title'], PDO::PARAM_STR);
+    $stmt->bindValue(':content', $post['content'], PDO::PARAM_STR);
+    $stmt->bindValue(':date', $post['date'], PDO::PARAM_STR);
+    $stmt->execute();
+}
+
 
 ?>
